pyproject.toml
[project]
name = "autonomous-writer"
version = "0.1.0"
description = "Autonomous AI blogging agent"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "tavily-python>=0.5",
    "atproto==0.0.65",
    "cryptography>=44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.15.8",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "ANN",   # flake8-annotations
    "RUF",   # ruff-specific
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN"]  # don't require annotations in tests
